+++ node/resolvers/postResolvers.ts
@@ -1,11 +1,15 @@
 /* eslint-disable @typescript-eslint/camelcase */
 export const postResolvers = {
-  author: async ({ author }: { author: number }, _: any, ctx: Context) => {
+  author: async (
+    { author }: { author: number },
+    { customDomain }: { customDomain: string },
+    ctx: Context
+  ) => {
     const {
       clients: { wordpressProxy },
     } = ctx
     try {
-      return await wordpressProxy.getUser(author)
+      return await wordpressProxy.getUser(author, customDomain)
     } catch (e) {
       console.error(e)
     }
@@ -13,23 +17,27 @@ export const postResolvers = {
   },
   categories: async (
     { categories }: { categories: [number] },
-    _: any,
+    { customDomain }: { customDomain: string },
     ctx: Context
   ) => {
     const {
       clients: { wordpressProxy },
     } = ctx
-    return categories.map(id => wordpressProxy.getCategory(id))
+    return categories.map(id => wordpressProxy.getCategory(id, customDomain))
   },
-  tags: async ({ tags }: { tags: [number] }, _: any, ctx: Context) => {
+  tags: async (
+    { tags }: { tags: [number] },
+    { customDomain }: { customDomain: string },
+    ctx: Context
+  ) => {
     const {
       clients: { wordpressProxy },
     } = ctx
-    return tags.map(id => wordpressProxy.getTag(id))
+    return tags.map(id => wordpressProxy.getTag(id, customDomain))
   },
   featured_media: async (
     { featured_media }: { featured_media: number },
-    _: any,
+    { customDomain }: { customDomain: string },
     ctx: Context
   ) => {
     const {
@@ -37,7 +45,7 @@ export const postResolvers = {
     } = ctx
     if (featured_media > 0) {
       try {
-        return await wordpressProxy.getMediaSingle(featured_media)
+        return await wordpressProxy.getMediaSingle(featured_media, customDomain)
       } catch (e) {
         console.error(e)
       }