+++ react/components/WordpressBreadcrumb.tsx
@@ -1,6 +1,7 @@
 import { Container } from 'vtex.store-components'
 
 import React, { FunctionComponent, Fragment } from 'react'
+import { defineMessages } from 'react-intl'
 import { useCssHandles } from 'vtex.css-handles'
 import { Link } from 'vtex.render-runtime'
 import { useQuery } from 'react-apollo'
@@ -10,14 +11,18 @@ import CategorySimpleBySlug from '../graphql/CategorySimpleBySlug.graphql'
 
 interface Props {
   params: any
+  customDomains: string
 }
 
 interface CategoryProps {
   categorySlug?: string
+  customDomain?: string
 }
 
 interface SinglePostProps {
   slug?: string
+  customDomain?: string
+  customDomainSlug?: string
 }
 
 const CSS_HANDLES = [
@@ -28,10 +33,13 @@ const CSS_HANDLES = [
   'breadcrumbCurrentPage',
 ] as const
 
-const WordpressCategoryBreadcrumb: FunctionComponent<CategoryProps> = props => {
+const WordpressCategoryBreadcrumb: FunctionComponent<CategoryProps> = ({
+  categorySlug,
+  customDomain,
+}) => {
   const handles = useCssHandles(CSS_HANDLES)
   const { data, loading, error } = useQuery(CategorySimpleBySlug, {
-    variables: { categorySlug: props.categorySlug },
+    variables: { categorySlug, customDomain },
   })
   if (loading || error) return <Fragment></Fragment>
   if (data?.wpCategories?.categories?.length)
@@ -53,10 +61,14 @@ const WordpressCategoryBreadcrumb: FunctionComponent<CategoryProps> = props => {
   return null
 }
 
-const WordpressSinglePostBreadcrumb: FunctionComponent<SinglePostProps> = props => {
+const WordpressSinglePostBreadcrumb: FunctionComponent<SinglePostProps> = ({
+  slug,
+  customDomain,
+  customDomainSlug,
+}) => {
   const handles = useCssHandles(CSS_HANDLES)
   const { data, loading, error } = useQuery(PostSimpleBySlug, {
-    variables: { slug: props.slug },
+    variables: { slug, customDomain },
   })
 
   if (loading || error) return <Fragment></Fragment>
@@ -76,6 +88,7 @@ const WordpressSinglePostBreadcrumb: FunctionComponent<SinglePostProps> = props 
           params={{
             categoryslug: data.wpPosts.posts[0].categories[0].slug,
             page: '1',
+            customdomainslug: customDomainSlug,
           }}
           className={`${handles.breadcrumbLink}`}
         >
@@ -90,18 +103,36 @@ const WordpressSinglePostBreadcrumb: FunctionComponent<SinglePostProps> = props 
   return null
 }
 
-const WordpressBreadcrumb: FunctionComponent<Props> = props => {
+const WordpressBreadcrumb: StorefrontFunctionComponent<Props> = ({
+  customDomains,
+  params,
+}) => {
   const handles = useCssHandles(CSS_HANDLES)
 
+  let parsedCustomDomains = null
+  try {
+    parsedCustomDomains = customDomains ? JSON.parse(customDomains) : null
+  } catch (e) {
+    console.error(e)
+  }
+
+  const customDomain =
+    params.customdomainslug && parsedCustomDomains
+      ? parsedCustomDomains[params.customdomainslug]
+      : undefined
+
   // if we're on a category page with a slug
-  if (props.params?.categoryslug) {
+  if (params?.categoryslug) {
     return (
-      <WordpressCategoryBreadcrumb categorySlug={props.params.categoryslug} />
+      <WordpressCategoryBreadcrumb
+        categorySlug={params.categoryslug}
+        customDomain={customDomain}
+      />
     )
   }
 
   // if we're on an article search page
-  if (props.params?.term) {
+  if (params?.term) {
     return (
       <Container
         className={`${handles.breadcrumbContainer} pt2 pb8`}
@@ -109,22 +140,31 @@ const WordpressBreadcrumb: FunctionComponent<Props> = props => {
       >
         <Link
           page="store.blog-home"
-          params={{ page: '1' }}
+          params={{
+            page: '1',
+            customdomainslug: params.customdomainslug,
+          }}
           className={`${handles.breadcrumbHomeLink}`}
         >
           Blog Home
         </Link>
         <span className={`${handles.breadcrumbSeparator}`}>&nbsp;/&nbsp;</span>
         <span className={`${handles.breadcrumbCurrentPage}`}>
-          Search results for &quot;{props.params.term}&quot;
+          Search results for &quot;{params.term}&quot;
         </span>
       </Container>
     )
   }
 
   // if we're viewing a single blog post with a slug
-  if (props.params?.slug) {
-    return <WordpressSinglePostBreadcrumb slug={props.params.slug} />
+  if (params?.slug) {
+    return (
+      <WordpressSinglePostBreadcrumb
+        slug={params.slug}
+        customDomain={customDomain}
+        customDomainSlug={params.customdomainslug}
+      />
+    )
   }
 
   // else
@@ -139,6 +179,44 @@ const WordpressBreadcrumb: FunctionComponent<Props> = props => {
       </Link>
     </Container>
   )
+}
+
+const messages = defineMessages({
+  title: {
+    defaultMessage: '',
+    id: 'admin/editor.wordpressBreadcrumb.title',
+  },
+  description: {
+    defaultMessage: '',
+    id: 'admin/editor.wordpressBreadcrumb.description',
+  },
+  customDomainsTitle: {
+    defaultMessage: '',
+    id: 'admin/editor.wordpressCustomDomains.title',
+  },
+  customDomainsDescription: {
+    defaultMessage: '',
+    id: 'admin/editor.wordpressCustomDomains.description',
+  },
+})
+
+WordpressBreadcrumb.defaultProps = {
+  customDomains: undefined,
+}
+
+WordpressBreadcrumb.schema = {
+  title: messages.title.id,
+  description: messages.description.id,
+  type: 'object',
+  properties: {
+    customDomains: {
+      title: messages.customDomainsTitle.id,
+      description: messages.customDomainsDescription.id,
+      type: 'string',
+      isLayout: false,
+      default: '',
+    },
+  },
 }
 
 export default WordpressBreadcrumb